moved classes into single file
