--- admin/error-types/mtc-base.error.js
@@ -1,9 +1,0 @@
-class mtcBaseError extends Error {
-  constructor (message, userMessage = '') {
-    super(message)
-    this.name = 'mtcBaseError'
-    this.userMessage = userMessage
-  }
-}
-
-module.exports = mtcBaseError
