+++ admin/services/nca-tools-user.service.js
@@ -3,7 +3,7 @@
 const roleService = require('./role.service')
 const schoolDataService = require('./data-access/school.data.service')
 const userDataService = require('./data-access/user.data.service')
-const MtcHelpdeskImpersonationError = require('../error-types/mtc-helpdesk-impersonation.error')
+const { MtcHelpdeskImpersonationError } = require('../error-types/mtc-error')
 
 const service = {
   /**
