+++ admin/spec/back-end/service/check-window-v2.service.spec.js
@@ -6,7 +6,7 @@ const uuid = require('uuid/v4')
 const checkWindowDataService = require('../../../services/data-access/check-window.data.service')
 const checkWindowV2Service = require('../../../services/check-window-v2.service')
 const dateService = require('../../../services/date.service')
-const MtcCheckWindowNotFoundError = require('../../../error-types/mtc-check-window-not-found.error')
+const { MtcCheckWindowNotFoundError } = require('../../../error-types/mtc-error')
 
 describe('check-window-v2.service', () => {
   describe('getCheckWindow', () => {
